Add database seeder and configure default parking spot data
- Created ParkingSeeder to populate the database with initial parking spot data (10 spots)
- Registered AppDbContext in the DI container
- Added seeding logic to Program.cs to run at application startup
- Changed ParkingSpotSize properties from float to double

+++ Parking/Parking/Entities/ParkingSpotSize.cs
@@ -2,8 +2,8 @@
 {
     public class ParkingSpotSize
     {
-        public float Width { get; set; }
-        public float Length { get; set; }
-        public float MaxVehicleHeight { get; set; }
+        public double Width { get; set; }
+        public double Length { get; set; }
+        public double MaxVehicleHeight { get; set; }
     }
 }
